Extract generateSpiMessage into its own file
And then update imports
