+++ characterisation/exceptions/HO100108.test.ts
@@ -1,8 +1,8 @@
 jest.setTimeout(30000)
 
 import World from "../../utils/world"
-import type { GenerateSpiMessageOptions } from "../helpers/generateMessage"
-import { generateSpiMessage } from "../helpers/generateMessage"
+import type { GenerateSpiMessageOptions } from "../helpers/generateSpiMessage"
+import generateSpiMessage from "../helpers/generateSpiMessage"
 import { processPhase1Message } from "../helpers/processMessage"
 
 describe.ifPhase1("HO100108", () => {
