+++ characterisation/exceptions/HO100236.test.ts
@@ -2,7 +2,7 @@ jest.setTimeout(30000)
 
 import World from "../../utils/world"
 import { lookupOffenceByCjsCode } from "../helpers/dataLookup"
-import { generateSpiMessage } from "../helpers/generateMessage"
+import generateSpiMessage from "../helpers/generateSpiMessage"
 import { processPhase1Message } from "../helpers/processMessage"
 
 jest.mock("../helpers/dataLookup", () => ({
