+++ characterisation/triggers/TRPR0008.test.ts
@@ -1,7 +1,7 @@
 jest.setTimeout(30000)
 
 import World from "../../utils/world"
-import { generateSpiMessage } from "../helpers/generateMessage"
+import generateSpiMessage from "../helpers/generateSpiMessage"
 import { processPhase1Message } from "../helpers/processMessage"
 import { SpiPlea } from "../types/Plea"
 import { TriggerCode } from "../types/TriggerCode"
